Add logic to hide edit/delete button
Some notes…

Beginner Reminder: in the if statement, currentUser will be evaluated
first. If its empty then the statement short circuits. There must be
something in the var for the second statement to evaluate. If its empty
and put through the function equals(), the app will break.

The variable currentUser comes from app.js when we first set up
authorization. The included comment points to the statement that makes
it available for the whole app. That statement is within middleware
that is used in every route.